+++ backend/src/main/java/com/workout/body/dto/BodyCompositionResponse.java
@@ -1,14 +1,16 @@
 package com.workout.body.dto;
 
 import com.workout.body.domain.BodyComposition;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record BodyCompositionResponse(
     Long id,
     LocalDate measurementDate,
-    Long weightKg,
-    Long fatKg,
-    Long muscleMassKg,
+    BigDecimal weightKg,
+    BigDecimal fatKg,
+    BigDecimal muscleMassKg,
     Long memberId
 ) {
   public static BodyCompositionResponse from(BodyComposition bodyComposition) {
